[Placement Group] Fix infeasible placement group not scheduled after node is added (#21993)
It looks like existing infeasible placement group in placement group manager didn't work properly. Idk how we added this feature when we cannot pass this simple test case.

But this is what has happend;

(1) PG is not scheduleable because it is infeasible
(2) New node is added
(3) After a new node is added, placement group manager tries rescheduling all infeasible pgs.
(4) Here, when we add a new node, we didn't report resources (this seems to be very weird. We are reporting resource using a separate RPC here). So when (3) happens, pg was still unschedulable. 

This PR fixes the issue by adding the resource information when the new node is added. 

Note that in the long term, we'd like to have a separate resource path from (4). This won't be addressed in this PR.