For a dataset comprised of both empty and non-empty blocks, let the non-empty blocks determine the schema (#22834)
There is a bug in combining the results from map_batches: if we create two dataset out of the same data, but with different num of partitions, we may get different results when run the same map_batches() on them. That is, num of partitions is affecting the map_batches() results, which should not.