[serve] Make deployments in `Application` use only import paths (#23027)
`Application` stores a group of deployments and can write them to a YAML config. However, this requires the deployments to use import paths as their `func_or_class`. This change make all deployments in an `Application` store only import paths as the `func_or_class`.

This change also adds a utility function to get a deployment's import path. This utility function is used in the DeploymentNode for Pipelines.