[Core][Java][Remove JVM FullGC 3/N]  Disable every 10min FullGC. (#21443)
In this PR, we disabled every 10min FullGC which is not triggered by a global gc event in Java worker. As detail, we added `triggered_by_global_gc` flag to indicate whether the gc event is triggered by a global gc event. If it's triggered by global gc, we still need to do FullGC.

Co-authored-by: Qing Wang <jovany.wq@antgroup.com>