Migrating to flat hash map [core worker&object manager] (#23126)
Next move of #22932. This pr replace unordered_map to flat_hash_map in core worker and object manager module.
Also some interfaces, like GetAllReferenceCounts, which expose user interfaces in Java/Python, is exclusive as it's a little bit complicated. We save them to deal with pg together.

The follow-up PRs would be migrating in reference counting, placement group and others.